+++ src/test/groovy/com/oroplatform/idea/oroplatform/intellij/codeAssist/yml/v1/SearchCompletionTest.groovy
@@ -63,7 +63,7 @@ class SearchCompletionTest extends CompletionTest {
             |    -
             |      target_type: <caret>
             """.stripMargin(),
-            ["text", "integer", "double", "datetime"]
+            ["text", "integer", "decimal", "datetime"]
         )
     }
